Only the fly index creation for testcsv
