+++ lib/testcsv.js
@@ -78,7 +78,7 @@ function test(argv, cb) {
         index: argv.index
     }
 
-    const tokens_raw = carmencnf.getInfo.metadata.geocoder_tokens;
+    const tokens_raw = carmencnf.getInfo.metadata ? carmencnf.getInfo.metadata.geocoder_tokens : carmencnf.getInfo.geocoder_tokens;
     const tokens = {};
 
     for (let token in tokens_raw) {
