+++ test/fixtures/test-ri/carmen-config2.json
@@ -38,7 +38,7 @@
         "geocoder_address": 1,
         "geocoder_format": "{address._number} {address._name}, {locality._name}, {place._name}, {region._name} {postcode._name}, {country._name}",
         "geocoder_resolution": 0,
-        "geocoder_version": 8,
+        "geocoder_version": 9,
         "geocoder_stack": [
             "us"
         ],
