--- test/fixtures/test-ri/index/us_ri-address-both-0d603c2a171017011038-0d603c2a39.freq.rocksdb/CURRENT
@@ -1,1 +1,0 @@
-MANIFEST-000001
