--- test/fixtures/test-ri/index/us_ri-address-both-0d603c2a171017011038-0d603c2a39.freq.rocksdb/IDENTITY
@@ -1,1 +1,0 @@
-72c3a674-08dd-411f-af0d-d5967b28d1e6
