--- test/fixtures/test-ri/index/us_ri-address-both-0d603c2a171017011038-0d603c2a39.grid.rocksdb/IDENTITY
@@ -1,1 +1,0 @@
-2c0c755f-fabc-4f9a-a606-5efd8d7c26d5
