--- test/fixtures/test-ri/index/us_ri-address-both-0d603c2a171017011038-0d603c2a39.norm.rocksdb/IDENTITY
@@ -1,1 +1,0 @@
-87398b3d-59e6-46f1-8082-9d63bd7160f6
