--- test/fixtures/test-ri/index/us_ri-address-both-0d603c2a171017011038-0d603c2a39.norm.rocksdb/OPTIONS-000005
@@ -1,135 +1,0 @@
-# This is a RocksDB option file.
-#
-# For detailed file format spec, please refer to the example file
-# in examples/rocksdb_option_file_example.ini
-#
-
-[Version]
-  rocksdb_version=4.13.4
-  options_file_version=1.1
-
-[DBOptions]
-  info_log_level=INFO_LEVEL
-  write_thread_max_yield_usec=100
-  write_thread_slow_yield_usec=3
-  fail_if_options_file_error=false
-  stats_dump_period_sec=600
-  max_total_wal_size=0
-  wal_recovery_mode=kPointInTimeRecovery
-  wal_bytes_per_sync=0
-  max_manifest_file_size=18446744073709551615
-  bytes_per_sync=0
-  max_subcompactions=1
-  WAL_ttl_seconds=0
-  wal_dir=/tmp/temp.adhfo
-  enable_write_thread_adaptive_yield=false
-  manifest_preallocation_size=4194304
-  log_file_time_to_roll=0
-  recycle_log_file_num=0
-  allow_concurrent_memtable_write=false
-  keep_log_file_num=1000
-  db_write_buffer_size=0
-  dump_malloc_stats=false
-  table_cache_numshardbits=6
-  max_open_files=-1
-  base_background_compactions=1
-  max_background_compactions=1
-  db_log_dir=
-  use_fsync=false
-  use_adaptive_mutex=false
-  writable_file_max_buffer_size=1048576
-  max_background_flushes=1
-  compaction_readahead_size=0
-  skip_stats_update_on_db_open=false
-  skip_log_error_on_recovery=false
-  new_table_reader_for_compaction_inputs=false
-  error_if_exists=false
-  enable_thread_tracking=false
-  delete_obsolete_files_period_micros=21600000000
-  random_access_max_buffer_size=1048576
-  is_fd_close_on_exec=true
-  disable_data_sync=false
-  WAL_size_limit_MB=0
-  create_missing_column_families=false
-  paranoid_checks=true
-  disableDataSync=false
-  create_if_missing=true
-  access_hint_on_compaction_start=NORMAL
-  max_log_file_size=0
-  allow_2pc=false
-  allow_mmap_writes=false
-  delayed_write_rate=2097152
-  max_file_opening_threads=16
-  allow_fallocate=true
-  avoid_flush_during_recovery=false
-  allow_os_buffer=true
-  allow_mmap_reads=false
-  advise_random_on_open=true
-  
-
-[CFOptions "default"]
-  compaction_filter_factory=nullptr
-  memtable_factory=SkipListFactory
-  bottommost_compression=kDisableCompressionOption
-  compression=kNoCompression
-  max_sequential_skip_in_iterations=8
-  max_bytes_for_level_multiplier_additional=1:1:1:1:1:1:1
-  compression_per_level=
-  max_bytes_for_level_multiplier=10
-  max_bytes_for_level_base=268435456
-  min_partial_merge_operands=2
-  table_factory=BlockBasedTable
-  max_successive_merges=0
-  arena_block_size=8388608
-  merge_operator=nullptr
-  target_file_size_multiplier=1
-  num_levels=7
-  min_write_buffer_number_to_merge=1
-  prefix_extractor=nullptr
-  bloom_locality=0
-  max_write_buffer_number=2
-  level0_stop_writes_trigger=24
-  level0_slowdown_writes_trigger=20
-  level0_file_num_compaction_trigger=4
-  write_buffer_size=67108864
-  memtable_huge_page_size=0
-  max_compaction_bytes=1677721600
-  hard_pending_compaction_bytes_limit=274877906944
-  target_file_size_base=67108864
-  soft_pending_compaction_bytes_limit=68719476736
-  comparator=UInt32Comparator
-  verify_checksums_in_compaction=true
-  force_consistency_checks=false
-  max_write_buffer_number_to_maintain=0
-  paranoid_file_checks=false
-  optimize_filters_for_hits=false
-  level_compaction_dynamic_level_bytes=false
-  purge_redundant_kvs_while_flush=true
-  inplace_update_support=false
-  compaction_style=kCompactionStyleLevel
-  compaction_filter=nullptr
-  disable_auto_compactions=false
-  inplace_update_num_locks=10000
-  memtable_prefix_bloom_size_ratio=0.000000
-  report_bg_io_stats=false
-  
-[TableOptions/BlockBasedTable "default"]
-  read_amp_bytes_per_bit=8589934592
-  skip_table_builder_flush=false
-  filter_policy=nullptr
-  verify_compression=false
-  block_restart_interval=16
-  block_size_deviation=10
-  format_version=2
-  index_block_restart_interval=1
-  block_size=4096
-  whole_key_filtering=true
-  no_block_cache=false
-  checksum=kCRC32c
-  hash_index_allow_collision=true
-  index_type=kBinarySearch
-  pin_l0_filter_and_index_blocks_in_cache=false
-  cache_index_and_filter_blocks_with_high_priority=false
-  cache_index_and_filter_blocks=false
-  flush_block_policy_factory=FlushBlockBySizePolicyFactory
-
